Create a dynamic base chart component and delete other existing component
Keep constant with type of chart for future purpose
Create a reset chart to force redraw of new chart at the next micro task
Store a reference to the canvas to ease the redraw
Make the reactivityManager ask for a micro task if the underlying property has changed
Create a type attribut with reactivity and bind its value to aria label if no title is present for accessibility
