--- force-app/main/default/lwc/barChart/barChart.html
@@ -1,10 +1,0 @@
-<template>
-  <div
-    style={chartStyle}
-    lwc:dom="manual"
-    role="img"
-    aria-label={ariaLabel}
-  ></div>
-
-  <slot></slot>
-</template>
