--- force-app/main/default/lwc/barChart/barChart.js
@@ -1,6 +1,0 @@
-import BaseChart from 'c/baseChart';
-import { BAR_CHART_TYPE } from 'c/constants';
-
-export default class BarChart extends BaseChart {
-  static type = BAR_CHART_TYPE;
-}
