--- force-app/main/default/lwc/bubbleChart/bubbleChart.css
@@ -1,3 +1,0 @@
-:host {
-  display: block;
-}
