--- force-app/main/default/lwc/bubbleChart/bubbleChart.js
@@ -1,5 +1,0 @@
-import BaseChart from 'c/baseChart';
-import { BUBBLE_CHART_TYPE } from 'c/constants';
-export default class BubbleChart extends BaseChart {
-  static type = BUBBLE_CHART_TYPE;
-}
