--- force-app/main/default/lwc/bubbleChart/bubbleChart.js-meta.xml
@@ -1,5 +1,0 @@
-<?xml version="1.0" encoding="UTF-8" ?>
-<LightningComponentBundle xmlns="http://soap.sforce.com/2006/04/metadata">
-    <apiVersion>48.0</apiVersion>
-    <isExposed>true</isExposed>
-</LightningComponentBundle>
