+++ force-app/main/default/lwc/chart/chart.html
