--- force-app/main/default/lwc/doughnutChart/doughnutChart.css
@@ -1,3 +1,0 @@
-:host {
-  display: block;
-}
