--- force-app/main/default/lwc/doughnutChart/doughnutChart.html
@@ -1,9 +1,0 @@
-<template>
-  <div
-    style={chartStyle}
-    lwc:dom="manual"
-    role="img"
-    aria-label={ariaLabel}
-  ></div>
-  <slot></slot>
-</template>
