--- force-app/main/default/lwc/doughnutChart/doughnutChart.js
@@ -1,5 +1,0 @@
-import BaseChart from 'c/baseChart';
-import { DOUGHNUT_CHART_TYPE } from 'c/constants';
-export default class DoughnutChart extends BaseChart {
-  static type = DOUGHNUT_CHART_TYPE;
-}
