--- force-app/main/default/lwc/dynamicChart/dynamicChart.css
@@ -1,3 +1,0 @@
-:host {
-  display: block;
-}
