--- force-app/main/default/lwc/lineChart/lineChart.css
@@ -1,3 +1,0 @@
-:host {
-  display: block;
-}
