--- force-app/main/default/lwc/lineChart/lineChart.html
@@ -1,9 +1,0 @@
-<template>
-  <div
-    style={chartStyle}
-    lwc:dom="manual"
-    role="img"
-    aria-label={ariaLabel}
-  ></div>
-  <slot></slot>
-</template>
