--- force-app/main/default/lwc/lineChart/lineChart.js
@@ -1,5 +1,0 @@
-import BaseChart from 'c/baseChart';
-import { LINE_CHART_TYPE } from 'c/constants';
-export default class LineChart extends BaseChart {
-  static type = LINE_CHART_TYPE;
-}
