--- force-app/main/default/lwc/pieChart/pieChart.css
@@ -1,3 +1,0 @@
-:host {
-  display: block;
-}
