--- force-app/main/default/lwc/pieChart/pieChart.js
@@ -1,5 +1,0 @@
-import BaseChart from 'c/baseChart';
-import { PIE_CHART_TYPE } from 'c/constants';
-export default class PieChart extends BaseChart {
-  static type = PIE_CHART_TYPE;
-}
