--- force-app/main/default/lwc/polarAreaChart/polarAreaChart.css
@@ -1,3 +1,0 @@
-:host {
-  display: block;
-}
