--- force-app/main/default/lwc/polarAreaChart/polarAreaChart.js
@@ -1,8 +1,0 @@
-import BaseChart from 'c/baseChart';
-import { POLARAREA_CHART_TYPE } from 'c/constants';
-/**
- * https://www.chartjs.org/docs/latest/charts/polar.html
- */
-export default class PolarAreaChart extends BaseChart {
-  static type = POLARAREA_CHART_TYPE;
-}
