--- force-app/main/default/lwc/radarChart/radarChart.css
@@ -1,3 +1,0 @@
-:host {
-  display: block;
-}
