--- force-app/main/default/lwc/scatterChart/scatterChart.css
@@ -1,3 +1,0 @@
-:host {
-  display: block;
-}
