--- force-app/main/default/lwc/scatterChart/scatterChart.html
@@ -1,9 +1,0 @@
-<template>
-  <div
-    style={chartStyle}
-    lwc:dom="manual"
-    role="img"
-    aria-label={ariaLabel}
-  ></div>
-  <slot></slot>
-</template>
