--- force-app/main/default/lwc/scatterChart/scatterChart.js
@@ -1,5 +1,0 @@
-import BaseChart from 'c/baseChart';
-import { SCATTER_CHART_TYPE } from 'c/constants';
-export default class ScatterChart extends BaseChart {
-  static type = SCATTER_CHART_TYPE;
-}
